requirements.txt
numpy == 2.0.1
scipy == 1.14.0
future == 1.0.0
matplotlib == 3.9.2
pillow == 10.4.0
opencv-python == 4.10.0.84
scikit-image == 0.24.0
pyaml == 24.7.0
streamlit == 1.37.1
torch == 2.4.0
boto3 == 1.34.162
